+++ build_cli/CHANGELOG.md
@@ -1,10 +1,11 @@
 ## 1.3.0
 
+- Added support for the new `provideDefaultToOverride` value on `CliOption` to
+  allow users to provide an override to `defaultsTo` by including an optional
+  `[fieldName]DefaultOverride` param to the generated 
+  `_$populate[ClassName]Parser` function.
 - Handle kebab-case enum value names correctly.
 - Handle allowed values that require escaping.
-
-## 1.2.3
-
 - Require Dart `>=2.2.0 <3.0.0`.
 
 ## 1.2.2